requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic>=2.10.3
pydantic-settings>=2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
qrcode[pil]==8.0
pandas==2.3.3
openpyxl==3.1.5
python-dotenv==1.0.1
alembic==1.14.0
